handler/requirements.txt
Click==7.0
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
Flask==1.1.1
Jinja2==2.10.1
Mako==1.0.13
Markdown==3.1.1
MarkupSafe==1.1.1
PyChromeDevTools==0.1
PyYAML==5.1.1
SQLAlchemy==1.3.5
SecretStorage==2.3.1
WTForms==2.2.1
Werkzeug==0.15.4
absl-py==0.7.1
alembic==1.0.11
asn1crypto==0.24.0
astor==0.8.0
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
beautifulsoup4==4.7.1
bleach==1.5.0
cffi==1.12.3
cryptography==2.1.4
cycler==0.10.0
dnspython==1.16.0
enum34==1.1.6
funcsigs==1.0.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
html5lib==0.9999999
idna==2.6
ipaddr==2.2.0
ipaddress==1.0.17
ipwhois==1.1.0
itsdangerous==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
mock==3.0.5
protobuf==3.8.0
pycparser==2.19
pycrypto==2.6.1
pyparsing==2.4.0
python-Levenshtein==0.12.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
pyxdg==0.25
six==1.12.0
soupsieve==1.9.2
subprocess32==3.5.4
termcolor==1.1.0
tld==0.9.3
wrapt==1.11.2